Removed/consolidated Inspec overrides
