+++ provider/inspec.rb
@@ -15,8 +15,6 @@ require 'google/ruby_utils'
 require 'provider/config'
 require 'provider/core'
 require 'provider/inspec/manifest'
-require 'provider/inspec/resource_override'
-require 'provider/inspec/property_override'
 require 'provider/overrides/inspec/resource_override'
 require 'provider/overrides/inspec/property_override'
 require 'active_support/inflector'
